(HomeViewModel) : add functionality to send tokens

+++ TradeBro/Models/Network/TokenMetadata.swift
@@ -5,7 +5,7 @@ struct TokenMetadata: Decodable {
     let logoURI: String
     let symbol: String
     let name: String
-    let decimals: Int
+    let decimals: UInt8
     let tags: [String]
     let extensions: [String: String]
     let price: Double?
